Moved header tests to the appropriate file
