+++ Tests/test_PDB_parse_pdb_header.py
@@ -18,11 +18,50 @@ except ImportError:
         "Install NumPy if you want to use Bio.PDB."
     ) from None
 
+from Bio.PDB import PDBParser
 from Bio.PDB.parse_pdb_header import parse_pdb_header, _parse_remark_465
 
 
 class ParseReal(unittest.TestCase):
     """Testing with real PDB file(s)."""
+
+    def test_1(self):
+        """Parse the header of a known PDB file (1A8O)."""
+        parser = PDBParser()
+        struct = parser.get_structure("1A8O", "PDB/1A8O.pdb")
+        self.assertAlmostEqual(struct.header["resolution"], 1.7)
+        # Case-insensitive string comparisons
+        known_strings = {
+            "author": "T.R.Gamble,S.Yoo,F.F.Vajdos,U.K.Von Schwedler,D.K.Worthylake,H.Wang,J.P.Mccutcheon,W.I.Sundquist,C.P.Hill",
+            "deposition_date": "1998-03-27",
+            "head": "viral protein",
+            "journal": "AUTH   T.R.GAMBLE,S.YOO,F.F.VAJDOS,U.K.VON SCHWEDLER,AUTH 2 D.K.WORTHYLAKE,H.WANG,J.P.MCCUTCHEON,W.I.SUNDQUIST,AUTH 3 C.P.HILLTITL   STRUCTURE OF THE CARBOXYL-TERMINAL DIMERIZATIONTITL 2 DOMAIN OF THE HIV-1 CAPSID PROTEIN.REF    SCIENCE                       V. 278   849 1997REFN                   ISSN 0036-8075PMID   9346481DOI    10.1126/SCIENCE.278.5339.849",
+            "journal_reference": "t.r.gamble,s.yoo,f.f.vajdos,u.k.von schwedler, d.k.worthylake,h.wang,j.p.mccutcheon,w.i.sundquist, c.p.hill structure of the carboxyl-terminal dimerization domain of the hiv-1 capsid protein. science v. 278 849 1997 issn 0036-8075 9346481 10.1126/science.278.5339.849 ",
+            "keywords": "capsid, core protein, hiv, c-terminal domain, viral protein",
+            "name": "hiv capsid c-terminal domain",
+            "release_date": "1998-10-14",
+            "structure_method": "x-ray diffraction",
+        }
+        for key, expect in known_strings.items():
+            self.assertEqual(struct.header[key].lower(), expect.lower())
+
+    def test_2(self):
+        """Parse the header of another PDB file (2BEG)."""
+        parser = PDBParser()
+        struct = parser.get_structure("2BEG", "PDB/2BEG.pdb")
+        known_strings = {
+            "author": "T.Luhrs,C.Ritter,M.Adrian,D.Riek-Loher,B.Bohrmann,H.Dobeli,D.Schubert,R.Riek",
+            "deposition_date": "2005-10-24",
+            "head": "protein fibril",
+            "journal": "AUTH   T.LUHRS,C.RITTER,M.ADRIAN,D.RIEK-LOHER,B.BOHRMANN,AUTH 2 H.DOBELI,D.SCHUBERT,R.RIEKTITL   3D STRUCTURE OF ALZHEIMER'S AMYLOID-{BETA}(1-42)TITL 2 FIBRILS.REF    PROC.NATL.ACAD.SCI.USA        V. 102 17342 2005REFN                   ISSN 0027-8424PMID   16293696DOI    10.1073/PNAS.0506723102",
+            "journal_reference": "t.luhrs,c.ritter,m.adrian,d.riek-loher,b.bohrmann, h.dobeli,d.schubert,r.riek 3d structure of alzheimer's amyloid-{beta}(1-42) fibrils. proc.natl.acad.sci.usa v. 102 17342 2005 issn 0027-8424 16293696 10.1073/pnas.0506723102 ",
+            "keywords": "alzheimer's, fibril, protofilament, beta-sandwich, quenched hydrogen/deuterium exchange, pairwise mutagenesis, protein fibril",
+            "name": "3d structure of alzheimer's abeta(1-42) fibrils",
+            "release_date": "2005-11-22",
+            "structure_method": "solution nmr",
+        }
+        for key, expect in known_strings.items():
+            self.assertEqual(struct.header[key].lower(), expect.lower())
 
     def test_parse_pdb_with_remark_465(self):
         """Tests that parse_pdb_header now can identify some REMARK 465 entries."""
